backend/requirements.txt
anthropic>=0.17.0,<1.0.0
langchain>=0.1.10,<0.2.0
langchain-anthropic==0.1.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
pypdf==3.17.0
pytesseract==0.3.10
pillow==10.1.0
numpy==1.26.1
pandas==2.1.2
matplotlib==3.8.1
seaborn==0.13.0
scikit-learn==1.3.2
pytest==7.4.3
httpx==0.25.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic>=1.13.1,<2.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
flower==2.0.1
gunicorn==21.2.0
